+++ yeat/tests/__init__.py
@@ -7,8 +7,10 @@
 # Development Center.
 # -------------------------------------------------------------------------------------------------
 
+import json
 import multiprocessing
 import os
+from pathlib import Path
 from pkg_resources import resource_filename
 
 
@@ -20,3 +22,23 @@ def data_file(path):
 
 def get_core_count():
     return multiprocessing.cpu_count()
+
+
+def write_config(labels, wd, cfg):
+    data = json.load(open(data_file(f"configs/{cfg}")))
+    assemblers = []
+    for assembler in data["assemblers"]:
+        if assembler["label"] in labels:
+            assemblers.append(assembler)
+    data["assemblers"] = assemblers
+    json.dump(data, open(Path(wd) / cfg, "w"))
+    return assemblers
+
+
+def files_exists(wd, assemblers, expected):
+    analysis_dir = Path(wd).resolve() / "analysis"
+    for assembler in assemblers:
+        for sample in assembler["samples"]:
+            label = assembler["label"]
+            algorithm = assembler["algorithm"]
+            assert (analysis_dir / sample / label / algorithm / expected).exists()